Cambiando de float a Double. Opción 7 disponible para elegir una conversión diferente. Finalmente, se creó el método listarMonedas para mostrar todas las monedas disponibles, y el método seleccionarMoneda para seleccionar dinámicamente la moneda deseada.

+++ src/ConsultaApi.java
@@ -39,7 +39,7 @@ public class ConsultaApi {
         return new Gson().fromJson(response.body(), ConversionApi.class);
     }
 
-    public ConversionApi realizarConversion(String monedaBase, String monedaDestino, float valor) {
+    public ConversionApi realizarConversion(String monedaBase, String monedaDestino, double valor) {
         String direccion = this.URL_API + "/pair/%s/%s/%s"
                 .formatted(monedaBase, monedaDestino, String.valueOf(valor));
         HttpClient client = HttpClient.newHttpClient();
